use predetermined locks utils
