+++ test/bounty-contracts/prime-factoring-bounty/cost-of-solving-primes/cost-of-solving-primes-4608.test.ts
@@ -1,5 +1,5 @@
 import { BigNumber } from 'ethers'
-import { costOfSolvingPrimesFactory } from './cost-of-solving-primes/utilities'
+import { costOfSolvingPrimesFactory } from './cost-of-solving-primes-factory'
 
 const primesOf100 = [
   BigNumber.from(2),
@@ -24,7 +24,7 @@ const primes1536Bit = [
 ]
 primeFactors.concat(...primes1536Bit)
 
-describe.skip(
+describe(
   'Test the cost of solving the prime factoring bounty with 4608-bit key',
-  costOfSolvingPrimesFactory(128, primeFactors, 0x1b2befab, 0x4b423d)
+  costOfSolvingPrimesFactory(192, primeFactors, 0x1b2befab, 0x4b423d)
 )
