+++ test/bounty-contracts/prime-factoring-bounty/prime-factoring-bounty-with-predetermined-locks/prime-factoring-bounty-with-predetermined-locks-utils.ts
@@ -33,6 +33,11 @@ class PrimeFactoringBountyWithPredeterminedLocksUtils extends BountyUtils {
     }
   ]
 
+  constructor (lockAndKeys?: Array<{lock: string, keys: string[]}>) {
+    super()
+    if (lockAndKeys != null) this.locksAndKeys = lockAndKeys
+  }
+
   public async deployBounty (): Promise<PrimeFactoringBountyWithPredeterminedLocks> {
     const ethersSigner = ethers.provider.getSigner()
     const locks = await this.getLocks()
@@ -48,8 +53,12 @@ class PrimeFactoringBountyWithPredeterminedLocksUtils extends BountyUtils {
   }
 
   public async solveBounty (bounty: BountyContract, getUserBalance?: () => Promise<BigNumber>): Promise<SolveAttemptResult> {
-    const solutions = this._getPrimes().map(primes => this.encodeByteArray(primes))
+    const solutions = this.getSolutions()
     return solveBountyReturningUserBalanceBeforeFinalSolution(solutions, bounty, getUserBalance)
+  }
+
+  public getSolutions (): bytes[] {
+    return this._getPrimes().map(primes => this.encodeByteArray(primes))
   }
 
   public async solveBountyPartially (bounty: BountyContract): Promise<void> {
